Sample implementation of the new API Wrapper
Since the current LTS Version of nodejs (v6.9) has full support for classes the implementation makes heavy use of them.

The new implementation splits the lib into multiple files for a better overview and uses inheritance to keep it DRY.

Rate Information are kept and updated using the `X-Rate-Limit-Count` header.

Also to fight the rate limit it supports caching (currently no caching implemented)

Also it trys to make the components testable.

The old implementation did not have proper error handling, this branch trys to wrap around the errors from the api and give them to the user in a managed way.