refactor (Endpoint) * rename * request, caching and rate limit should be done within ApiRequest * apiUrl now is build on construction by Endpoint class and respective sub-classes * buildUrl is needed to build final url with platformId, query and Api-key
Endpoint class now serves as a generic Base-class for endpoints to construct
api-urls and delegate the final request.

The Endpoint specific actions will be done within the respective
implementations of this class (e.g. constructing the apiUrl
for the specific endpoint and delegating the final url construction
to the Endpoint class with appropriate parameters