refactor (Endpoints) BREAKING changes to signatures and platformId parameter
* switched Signature position of options <-> platformId parameters.
 method(id, platformId, options) is a bit more natural.
 Having the options object in the middle can make it hard to read and
 can more easily lead to errors due to forgotten platformId parameters.

* added EndpointUtil.extractPlatformIdAndOptions() returning {_platformId, _options}.
This method should be used in endpoint-methods that take an options object,
to make both parameters optional without having to provide a falsy value for one of them.
If both are provided, the order of the signature has to be respected though.

* methods now work with platformId and region-tag (case-insensitive).
Error is thrown if they are invalid.
Implementation note: for all Endpoints this check is done within the executing methods
in Endpoint class. For methods that also take an options object,
EnpointUtil.extractPlatformIdAndOptions() should be used.